Move unit into metrics module (#564)
