+++ opentelemetry/src/core.rs
@@ -260,26 +260,3 @@ impl KeyValue {
         }
     }
 }
-
-/// Units denote underlying data units tracked by `Meter`s.
-#[derive(Clone, Default, Debug, PartialEq, Hash)]
-pub struct Unit(String);
-
-impl Unit {
-    /// Create a new `Unit` from an `Into<String>`
-    pub fn new<S: Into<String>>(value: S) -> Self {
-        Unit(value.into())
-    }
-
-    /// View unit as &str
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl AsRef<str> for Unit {
-    #[inline]
-    fn as_ref(&self) -> &str {
-        self.0.as_ref()
-    }
-}
