+++ opentelemetry/src/metrics/value_recorder.rs
@@ -1,9 +1,8 @@
 use crate::metrics::{
     sync_instrument::{SyncBoundInstrument, SyncInstrument},
-    Descriptor, InstrumentKind, Measurement, Meter, Number, NumberKind, Result,
+    Descriptor, InstrumentKind, Measurement, Meter, Number, NumberKind, Result, Unit,
 };
 use crate::KeyValue;
-use crate::Unit;
 use std::marker;
 
 /// ValueRecorder is a metric that records per-request non-additive values.
